# api/src/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Asistencias",
    "description": "Para las rutas que no son GET es necesario inculuir en el 'headers' de las peticiones, la clave 'Authorization': 'Bearer JWT'. El JWT es generado por el servidor y mostrado en la consola.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:8080/api"
    }
  ],
  "paths": {
    "/attendance": {
      "delete": {
        "operationId": "d22a53b3-423c-40cf-82a0-e7faa9bb0c24",
        "tags": [
          "attendance"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DELETE /attendance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "ccac6081-a8c8-4da9-b21c-ced75eb74bfb",
        "tags": [
          "attendance"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "POST /attendance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    },
                    "nombre": {
                      "type": "string"
                    },
                    "apellido": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "6f5d7d36-207a-4ded-b6c1-b7bc2279a8bd",
        "tags": [
          "attendance"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "inasistenciaId": {
                    "type": "string"
                  },
                  "estado": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PUT /attendance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/students": {
      "post": {
        "operationId": "adcd26b0-e969-4ef1-b3f8-e464d810f574",
        "tags": [
          "students"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "apellido": {
                    "type": "string"
                  },
                  "telefono_contacto": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "POST /students",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "df4b12e8-eb83-4a57-b822-6d8b40837919",
        "tags": [
          "students"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apellido",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "grupo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "GET /students",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "nombre": {
                        "type": "string"
                      },
                      "apellido": {
                        "type": "string"
                      },
                      "activo": {
                        "type": "string"
                      },
                      "telefono_contacto": {
                        "type": "string"
                      },
                      "grupo_id": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "c13e0a77-362b-43eb-953f-3b17511b4bbd",
        "tags": [
          "students"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DELETE /students",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "5d9b9600-bcb1-4863-95de-d540083c0f7f",
        "tags": [
          "students"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "apellido": {
                    "type": "string"
                  },
                  "telefono_contacto": {
                    "type": "string"
                  },
                  "id": {
                    "type": "string"
                  },
                  "grupo_id": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PUT /students",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}